:sparkles: Split current logic between abstract class and implementation
