+++ src/index.ts
@@ -1,1 +1,1 @@
-export * from './lib/connector';
+export * from './lib/connectors/magicAuthConnector';
