+++ src/lib/connectors/magicConnector.ts
@@ -1,0 +1,133 @@
+import {OAuthExtension, OAuthProvider} from '@magic-ext/oauth';
+
+import {
+  Chain,
+  Connector,
+  normalizeChainId,
+} from '@wagmi/core';
+import { ethers, Signer } from 'ethers';
+import { getAddress } from 'ethers/lib/utils';
+
+import { createModal } from '../modal/view';
+import {InstanceWithExtensions, SDKBase} from "@magic-sdk/provider";
+import {ConnectExtension} from "@magic-ext/connect";
+
+const IS_SERVER = typeof window === 'undefined';
+
+export interface BaseOptions {
+  apiKey: string;
+  accentColor?: string;
+  isDarkMode?: boolean;
+  customLogo?: string;
+  customHeaderText?: string;
+  enableEmailLogin?: boolean;
+  enableSMSLogin?: boolean;
+  oauthOptions?: {
+    providers: OAuthProvider[];
+    callbackUrl?: string;
+  };
+}
+
+interface UserDetails {
+  email: string;
+  phoneNumber: string;
+  oauthProvider: OAuthProvider;
+}
+
+export abstract class MagicConnector extends Connector {
+  ready = !IS_SERVER;
+
+  readonly id = 'magic';
+
+  readonly name = 'Magic';
+
+  provider: any;
+
+  isModalOpen = false;
+
+  magicOptions: BaseOptions;
+
+  oauthProviders: OAuthProvider[];
+
+  oauthCallbackUrl?: string;
+
+  protected constructor(config: { chains?: Chain[]; options: BaseOptions }) {
+    super(config);
+    this.magicOptions = config.options;
+    this.oauthProviders = config.options.oauthOptions?.providers || [];
+    this.oauthCallbackUrl = config.options.oauthOptions?.callbackUrl;
+  }
+
+  async getAccount(): Promise<string> {
+    const provider = new ethers.providers.Web3Provider(
+      await this.getProvider()
+    );
+    const signer = provider.getSigner();
+    const account = await signer.getAddress();
+    return account;
+  }
+
+  async getUserDetailsByForm(): Promise<UserDetails> {
+    const output: UserDetails = (await createModal({
+      accentColor: this.magicOptions.accentColor,
+      isDarkMode: this.magicOptions.isDarkMode,
+      customLogo: this.magicOptions.customLogo,
+      customHeaderText: this.magicOptions.customHeaderText,
+      isSMSLoginEnabled: this.magicOptions.enableSMSLogin,
+      isEmailLoginEnabled: this.magicOptions.enableEmailLogin ?? true,
+      oauthProviders: this.oauthProviders,
+    })) as UserDetails;
+
+    this.isModalOpen = false;
+    return output;
+  }
+
+  async getProvider() {
+    if (this.provider) {
+      return this.provider;
+    }
+    const magic = this.getMagicSDK();
+    this.provider = magic.rpcProvider;
+    return this.provider;
+  }
+
+  async getSigner(): Promise<Signer> {
+    const provider = new ethers.providers.Web3Provider(
+      await this.getProvider()
+    );
+    const signer = await provider.getSigner();
+    return signer;
+  }
+
+  async isAuthorized() {
+    try {
+      const account = await this.getAccount();
+      return !!account;
+    } catch {
+      return false;
+    }
+  }
+
+  protected onAccountsChanged(accounts: string[]): void {
+    if (accounts.length === 0) this.emit('disconnect');
+    else this.emit('change', { account: getAddress(accounts[0]) });
+  }
+
+  protected onChainChanged(chainId: string | number): void {
+    const id = normalizeChainId(chainId);
+    const unsupported = this.isChainUnsupported(id);
+    this.emit('change', { chain: { id, unsupported } });
+  }
+
+  protected onDisconnect(): void {
+    this.emit('disconnect');
+  }
+
+  async disconnect(): Promise<void> {
+    const magic = this.getMagicSDK();
+    await magic.user.logout();
+  }
+
+  abstract getMagicSDK(): InstanceWithExtensions<SDKBase, OAuthExtension[]> | InstanceWithExtensions<SDKBase, ConnectExtension[]>;
+
+}
